Refactor Game Controller
Incorporated logic to use a counter when going forward to track the correct number of moves a user is going in between their actual roll.

Create if conditionals to not send an option if the user is on the first or last spot and can no longer move forward or backward.

 Changes to be committed:
	modified:   server/game/gameController.js