Implements Exit Game Early
Sets up a button on controll button which allows users to click in order
to leave game early.  I incorporate a checkbox to confirm exit.

This emits a message to server.  The server re-emits to all users in room
and changes their views to allow people to move to home view.

On branch master
Your branch is up-to-date with 'origin/master'.
Changes to be committed:
modified:   public/board/board.html
modified:   public/board/board.js
modified:   server/server.js