Comments Out Functions That Were Used During Developement
These functions would normally delete all entries in the database and then create
a single entry for each table, which allowed us to develop the application with
an initial set of data.

We no longer need this so I commented it out.  This way any games created by users
will be preserved even after restarting the server.

On branch fix-startup-conditions
Changes to be committed:
modified:   server/db/db.js