Add flex abilities add acknowledge.component
Since Angular modifies the dom elements flex does not work directly,
we have to use their tool and API to get flex to work.  I added these:
@angular/flex-layout, @angular/cdk.

I moved the header with title to app.component and removed it from the
index.html page in preparation to make it its own component.  I would
like to make it a navbar giving abilities to nav back to home pages,
see list of players in game/universe, status of games, game progress,
etc.  I want to shift all the logic of makeing sure that a user is
logged in to this component so that the there is no duplication of code
(not sure if this is the correct way to do this).

I added an acknowledge.component that has all the credits and
disclaimers. This can be expanded and some of this should also be
kept in the repo license.txt file.

I am using flex on both the signin and acknowledge pages.

 On branch master
 Changes to be committed:
	modified:   frontend/package-lock.json
	modified:   frontend/package.json
	new file:   frontend/src/app/acknowledge/acknowledge.component.css
	new file:   frontend/src/app/acknowledge/acknowledge.component.html
	new file:   frontend/src/app/acknowledge/acknowledge.component.spec.ts
	new file:   frontend/src/app/acknowledge/acknowledge.component.ts
	modified:   frontend/src/app/app-routing.module.ts
	modified:   frontend/src/app/app.component.css
	modified:   frontend/src/app/app.component.html
	modified:   frontend/src/app/app.module.ts
	modified:   frontend/src/app/signin/signin.component.css
	modified:   frontend/src/app/signin/signin.component.html
	modified:   frontend/src/app/signin/signin.component.ts
	modified:   frontend/src/index.html
	modified:   frontend/src/styles.css