Change how we create gameId
We were counting the number of games in the collection and using that
as the gameId.  This presented problems when we delete a record then
there will be conflicts in the gameId.  I decided to use a mongo record
id as the gameId.  I create it manually when we add a new game and set
it as __id and gameId ( after stringifying ).

I also removed any display of gameId on the UI.

 On branch master
 Changes to be committed:
	modified:   frontend/src/app/home/home.component.html
	modified:   frontend/src/app/lobby/lobby.component.html
	modified:   frontend/src/app/resume-lobby/resume-lobby.component.html
	modified:   server/game/gameController.js
	modified:   server/game/gameModel.js