Remove secondary count of users in gamelobby.
We had a list of users and also a count of users in the game lobby
which was getting out of sync.  I removed the count and exposed the
users list directly to those functions and just count the list.  This
simplified the logic and removed an endpoint as well as one round trip
request to the server from each of the clients.

 On branch master
 Changes to be committed:
	modified:   frontend/src/app/game-factory.service.ts
	modified:   frontend/src/app/home/home.component.ts
	modified:   server/game/gameController.js
	modified:   server/routes.js
	modified:   server/server.js