Prune and update documentation.
I updated the main readme with current install and build instructions.
I also added acknowledgements and put links to the authors pages.
Deleted references to routes, contributing, style, etc...

I deleted documentation on our routes both frontend and backend.  We
can refer to the source code for clarification and the documentation
quickly goes out of sync with the backend and I did not want the
maintenance overhead - they were already outdated.

Also deleted documentation on contributing and style guides.  We dont
follow them so why keep them as part of our project.  This is open
source and if people want to contribute they will we can review as
appropriate.

 On branch master
 Changes to be committed:
	deleted:    CONTRIBUTING.md
	modified:   PRESS-RELEASE.md
	modified:   README.md
	deleted:    STYLE-GUIDE.md
	deleted:    angularRoutes.md
	modified:   frontend/src/app/acknowledge/acknowledge.component.html
	deleted:    serverRoutes.md