Add a pokedex page.
Right now this just shows all the pokemon we have in the DB.  We could
make this a searchable list and make it more effecient.  There is no
way to navigate to this page yet.  The url is /pokedex

Discovered that we have 'gold' pokemon too.  Had to add some css to
the pokemon-panel to accomodate.

 On branch master
 Changes to be committed:
	modified:   frontend/src/app/app-routing.module.ts
	modified:   frontend/src/app/app.module.ts
	new file:   frontend/src/app/general.service.spec.ts
	new file:   frontend/src/app/general.service.ts
	new file:   frontend/src/app/pokedex/pokedex.component.css
	new file:   frontend/src/app/pokedex/pokedex.component.html
	new file:   frontend/src/app/pokedex/pokedex.component.spec.ts
	new file:   frontend/src/app/pokedex/pokedex.component.ts
	modified:   frontend/src/app/pokemon-panel/pokemon-panel.component.css
	modified:   server/pokemon/pokemonController.js
	modified:   server/routes.js